Refactor component into the provider.
